+++ src/beta/client.rs
@@ -7,13 +7,10 @@ use std::sync::Arc;
 use crate::client::MunaClient;
 use crate::services::{PredictionService, PredictorService};
 use super::openai::OpenAIClient;
-use super::remote::{BetaPredictionService, RemotePredictionService};
 
 /// Client for incubating features.
 #[derive(Clone)]
 pub struct BetaClient {
-    /// Make remote predictions.
-    pub predictions: BetaPredictionService,
     /// OpenAI-compatible client.
     pub openai: OpenAIClient,
 }
@@ -21,13 +18,11 @@ pub struct BetaClient {
 impl BetaClient {
 
     pub fn new(
-        client: Arc<MunaClient>,
+        _: Arc<MunaClient>,
         predictors: PredictorService,
         predictions: PredictionService,
     ) -> Self {
-        let beta_predictions = BetaPredictionService::new(client.clone());
-        let remote_predictions = RemotePredictionService::new(client);
-        let openai = OpenAIClient::new(predictors, predictions, remote_predictions);
-        Self { predictions: beta_predictions, openai }
+        let openai = OpenAIClient::new(predictors, predictions);
+        Self { openai }
     }
 }