+++ src/beta/mod.rs
@@ -5,5 +5,4 @@
 
 mod client;
 pub mod openai;
-pub mod remote;
 pub use client::*;