+++ src/services/mod.rs
@@ -3,8 +3,10 @@
 *   Copyright © 2026 NatML Inc. All Rights Reserved.
 */
 
+mod local;
 mod prediction;
 mod predictor;
+mod remote;
 mod user;
 
 pub use prediction::*;